+++ R/ib_betareg.R
@@ -82,6 +82,7 @@ ib.betareg <- function(object, thetastart=NULL, control=list(...), ...){
   control1$H <- 1L
   linkinv <- object$link$mean$linkinv
 
+  # initial value
   diff <- rep(NA_real_, control$maxit)
 
   # Iterative bootstrap algorithm:
@@ -127,11 +128,15 @@ ib.betareg <- function(object, thetastart=NULL, control=list(...), ...){
     # initialize test
     if(!k) tt_old <- test_theta+1
 
-    # Stopping criteria
-    # "no more progress" :
-    # if(tt_old <= test_theta) {break} else {tt_old <- test_theta}
+    # Alternative stopping criteria, early stop :
+    if(control$early_stop){
+      if(tt_old <= test_theta){
+        warning("Algorithm stopped because the objective function does not reduce")
+        break
+      }
+    }
 
-    # "statistically flat progress curve" :
+    # Alternative stopping criteria, "statistically flat progress curve" :
     if(k > 10L){
       try1 <- diff[k:(k-10)]
       try2 <- k:(k-10)
@@ -150,6 +155,8 @@ ib.betareg <- function(object, thetastart=NULL, control=list(...), ...){
     # update theta
     t0 <- t1
   }
+  # warning for reaching max number of iterations
+  if(k>=control$maxit) warning("maximum number of iteration reached")
 
   # update betareg object
   # FIXME: update object$loglik, object$scoring, object$residuals,
@@ -168,15 +175,11 @@ ib.betareg <- function(object, thetastart=NULL, control=list(...), ...){
   #                                     mu, object$prior.weights, dev) + 2 * object$rank
 
   # additional metadata
-  ib_warn <- NULL
-  if(k>=control$maxit) ib_warn <- gettext("maximum number of iteration reached")
-  if(tt_old<=test_theta) ib_warn <- gettext("objective function does not reduce")
-  ib_extra <- list(
+ ib_extra <- list(
     iteration = k,
     of = sqrt(drop(crossprod(delta))),
     estimate = t0,
     test_theta = test_theta,
-    ib_warn = ib_warn,
     boot = tmp_pi)
 
   new("IbBetareg",