+++ NEWS.md
@@ -6,6 +6,8 @@
 - Enhance simulation for `glm` by allowing user-defined function.
 - Enhance `simulation` for negative binomial family. 
 - New feature by exporting the `simulation` method.
+- Enhance by adding the `betareg` family.
+- Enhance the stopping criteria for the IB algorithm.
 
 ## ib 0.1.0
 - First version for submission to CRAN. The package is made available on GitHub ([github.com/SMAC-Group/ib/](https://github.com/SMAC-Group/ib/))